backend/requirements.txt
Django>=4.2,<5
djangorestframework
psycopg2-binary
gunicorn
celery[redis]
redis
django-cors-headers
djangorestframework-simplejwt
django-filter
requests